requirements.txt
acme==1.21.0
Django==4.2.1
psycopg2-binary
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.5.0